Factor out Rofl app data loading to hooks
